add flags to page head
