+++ example/app/root.tsx
@@ -1,18 +1,27 @@
-import type { MetaFunction } from '@remix-run/node';
+import type { MetaFunction, LoaderFunction } from '@remix-run/node';
+import { useLoaderData } from '@remix-run/react';
 import { Links, LiveReload, Meta, Outlet, Scripts, ScrollRestoration } from '@remix-run/react';
-
+import { renderFlagsToString } from 'remix-sdk';
+import { createClient } from './ld.server'
 export const meta: MetaFunction = () => ({
   charset: 'utf-8',
   title: 'New Remix App',
   viewport: 'width=device-width,initial-scale=1',
 });
-
+export const loader: LoaderFunction = async () =>{
+  const serverId = await createClient()
+  return await renderFlagsToString({key: serverId});
+}
 export default function App() {
+  const flagBootstrap = useLoaderData();
   return (
     <html lang="en">
       <head>
         <Meta />
         <Links />
+        <script dangerouslySetInnerHTML={{
+          __html: flagBootstrap,
+        }}></script>
       </head>
       <body>
         <Outlet />
