+++ example/package.json
@@ -12,15 +12,18 @@
     "@remix-run/react": "^1.5.1",
     "@remix-run/serve": "^1.5.1",
     "launchdarkly-js-client-sdk": "^2.22.1",
+    "launchdarkly-node-server-sdk": "^6.4.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "remix-sdk": "file:../lib"
+    "remix-sdk": "file:../lib",
+    "uuid": "^8.3.2"
   },
   "devDependencies": {
     "@remix-run/dev": "^1.5.1",
     "@remix-run/eslint-config": "^1.5.1",
     "@types/react": "^17.0.45",
     "@types/react-dom": "^17.0.17",
+    "@types/uuid": "^8.3.4",
     "eslint": "^8.15.0",
     "typescript": "^4.6.4"
   },
